+++ src/Model/Entity/Page.php
@@ -28,8 +28,8 @@ use MeCms\ORM\PostAndPageEntity;
  * @property string $text
  * @property int $priority
  * @property bool $active
- * @property \Cake\I18n\Time $created
- * @property \Cake\I18n\Time $modified
+ * @property \Cake\I18n\FrozenTime $created
+ * @property \Cake\I18n\FrozenTime $modified
  * @property \MeCms\Model\Entity\PagesCategory $category
  */
 class Page extends PostAndPageEntity
@@ -47,6 +47,6 @@ class Page extends PostAndPageEntity
      */
     protected function _getUrl(): ?string
     {
-        return $this->has('slug') ? Router::url(['_name' => 'page', $this->get('slug')], true) : null;
+        return $this->hasValue('slug') ? Router::url(['_name' => 'page', $this->get('slug')], true) : null;
     }
 }