+++ src/Model/Entity/PostsCategory.php
@@ -28,8 +28,8 @@ use Cake\Routing\Router;
  * @property string $slug
  * @property string $description
  * @property int $post_count
- * @property \Cake\I18n\Time $created
- * @property \Cake\I18n\Time $modified
+ * @property \Cake\I18n\FrozenTime $created
+ * @property \Cake\I18n\FrozenTime $modified
  * @property \MeCms\Model\Entity\PostsCategory[] $child_posts_categories
  * @property \MeCms\Model\Entity\PostsCategory $parent_posts_category
  */
@@ -59,6 +59,6 @@ class PostsCategory extends Entity
      */
     protected function _getUrl(): ?string
     {
-        return $this->has('slug') ? Router::url(['_name' => 'postsCategory', $this->get('slug')], true) : null;
+        return $this->hasValue('slug') ? Router::url(['_name' => 'postsCategory', $this->get('slug')], true) : null;
     }
 }